Move debugger tests into subdirectory
